backend/requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.23
pillow>=9.0
